requirements.txt
langchain
langgraph
langchain-openai
pymongo
python-dotenv
pydantic
pytest
mongomock
